requirements.txt
numpy~=1.24.4
numpy-stl~=3.1.1